requirements.txt
opencv-python==4.8.1.78
tensorflow==2.14.0
face-recognition==1.3.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
scikit-learn==1.3.2
matplotlib==3.7.2
requests==2.31.0
geopy==2.4.1
folium==0.15.0
exifread==3.0.0
keras==2.14.0
torch==2.1.0
torchvision==0.16.0
dlib==19.24.2
imutils==0.5.4
flask==3.0.0
werkzeug==3.0.1